# refactor(*): faster check for a string starting with a substring